app-service/python-db-using-appservice/fastapi-project/requirements.txt
fastapi
uvicorn[standard]
gunicorn
psycopg[binary]  # The driver to talk to PostgreSQL
azure-identity   # For the "Passwordless" Managed Identity we discussed